refactor(SwiftExampleApp): centralize asset-lock status constants + plan doc sync
The AssetLockStatus discriminants (0/1/2/3 -> Built/Broadcast/
InstantSendLocked/ChainLocked) are protocol constants from the Rust
side. Until this commit four separate views each carried their own
copy of the case-block label mapping, and the >= 2 "is fundable"
threshold lived inline at every usage site. Consolidates into a
single PersistentAssetLock extension in the example app:

- statusLabel: protocol-discriminant -> human-readable string.
- canFundIdentity: statusRaw >= 2 (resume button gates on this).
- isVisibleAsResumable: statusRaw >= 1 (Resumable section surfaces this).
- shortOutPointDisplay: txid-prefix-plus-vout format used by every row.

The fundability/visibility predicates live on the AssetLockResumeRow
protocol so test fakes get them for free without an explicit
PersistentAssetLock instance.

Also:
- Delete unused `relativeDateString` and `assetLockStatusLabel` statics
  from CreateIdentityView (leftovers from the in-form picker removed in
  f466b7c4c7).
- Remove now-duplicate `statusLabel(_:)` helpers from
  StorageRecordDetailViews, StorageModelListViews, and IdentitiesContentView.
- Remove the duplicated `shortOutPoint(_:)` helper that lived in both
  CreateIdentityView and IdentitiesContentView.
- Sync the iter 5 prose in CREATE_IDENTITY_FROM_CORE_FUNDS_PLAN.md to
  reflect (a) the >= 1 visibility floor + active-controller anti-join
  and (b) the Identities-tab resume surface that replaced the original
  in-form picker.

Tests: 10/10 green, no behavior change.
